requirements.txt
numpy==1.24.4
scikit_learn==1.1.1
pytorch-crf==1.1.0
torchmetrics==1.4.0.post0
tqdm==4.64.1
transformers==4.30.2
